Proofreading
Proofreading of the vast majority of the item descriptions, scans, and dialogue.

I also set spicy jerky to be extremely well preserved because, well, it's jerky.

I'm a GitHub noob, so it ended up showing that all files were modified, so I just didn't commit anything not from dialgue, objects, items, or tiles.

Also, this used the version off the forums, oops.  The only code I touched was the lifespan multiplier of spicy jerky, but this may override any weapons nerfs you've done or the like.